# Improve d2exp_buffered_n headline perf.