--- a/ryu/d2fixed.c
+++ b/ryu/d2fixed.c
@@ -765,10 +765,17 @@ int d2exp_buffered_n(double d, uint32_t precision, char* result) {
   } else {
     result[index++] = '+';
   }
-  if (exp < 10) {
-    result[index++] = '0';
+
+  if (exp >= 100) {
+    const int32_t c = exp % 10;
+    memcpy(result + index, DIGIT_TABLE + 2 * (exp / 10), 2);
+    result[index + 2] = (char) ('0' + c);
+    index += 3;
+  } else {
+    memcpy(result + index, DIGIT_TABLE + 2 * exp, 2);
+    index += 2;
   }
-  index += append_n_digits(exp, result + index);
+
   return index;
 }